refactor: rename common directory to shared
