+++ src/app.module.ts
@@ -7,7 +7,7 @@ import { AccountModule } from './user/account/account.module';
 import { SignInModule } from './user/signin/signin.module';
 import { jwtConstants } from './user/utils';
 import { MovieModule } from './movie/movie.module';
-import { RedisModule } from './common/redis/redis.module';
+import { RedisModule } from './shared/redis/redis.module';
 
 @Module({
   imports: [
